# wfe-restapi-client/src/main/resources/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "runawfe api",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/restapi/executor/user": {
      "get": {
        "tags": [
          "executor-controller"
        ],
        "summary": "getUser",
        "operationId": "getUserUsingGET",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeUser"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "executor-controller"
        ],
        "summary": "createUser",
        "operationId": "createUserUsingPUT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfeUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeUser"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "executor-controller"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPATCH",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfeUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/restapi/task/delegate": {
      "patch": {
        "tags": [
          "task-controller"
        ],
        "summary": "delegateTasks",
        "operationId": "delegateTasksUsingPATCH",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "currentOwnerName",
            "in": "query",
            "description": "currentOwnerName",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keepCurrentOwner",
            "in": "query",
            "description": "keepCurrentOwner",
            "required": true,
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/list": {
      "post": {
        "tags": [
          "task-controller"
        ],
        "summary": "getTasks",
        "operationId": "getTasksUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfePagedListFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfePagedListOfWfeTask"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/my": {
      "post": {
        "tags": [
          "task-controller"
        ],
        "summary": "getMyTasks",
        "operationId": "getMyTasksUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfePagedListFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfePagedListOfWfeTask"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/process/{processId}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "summary": "getProcessTasks",
        "operationId": "getProcessTasksUsingGET",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "description": "processId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "includeSubprocesses",
            "in": "query",
            "description": "includeSubprocesses",
            "required": false,
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WfeTask"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/reassign": {
      "patch": {
        "tags": [
          "task-controller"
        ],
        "summary": "reassignTasks",
        "operationId": "reassignTasksUsingPATCH",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WfePagedListFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/unassigned": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "summary": "getUnassignedTasks",
        "operationId": "getUnassignedTasksUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WfeTask"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "summary": "getTask",
        "operationId": "getTaskUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeTask"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}/assign": {
      "patch": {
        "tags": [
          "task-controller"
        ],
        "summary": "assignTask",
        "operationId": "assignTaskUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "currentOwnerName",
            "in": "query",
            "description": "currentOwnerName",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newOwnerName",
            "in": "query",
            "description": "newOwnerName",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}/complete": {
      "post": {
        "tags": [
          "task-controller"
        ],
        "summary": "completeTask",
        "operationId": "completeTaskUsingPOST",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeTask"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeValidationExceptionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}/delegate": {
      "patch": {
        "tags": [
          "task-controller"
        ],
        "summary": "delegateTask",
        "operationId": "delegateTaskUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "currentOwnerName",
            "in": "query",
            "description": "currentOwnerName",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keepCurrentOwner",
            "in": "query",
            "description": "keepCurrentOwner",
            "required": true,
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}/markOpened": {
      "patch": {
        "tags": [
          "task-controller"
        ],
        "summary": "markTaskOpened",
        "operationId": "markTaskOpenedUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}/reassign": {
      "patch": {
        "tags": [
          "task-controller"
        ],
        "summary": "reassignTask",
        "operationId": "reassignTaskUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    },
    "/restapi/task/{id}/variable": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "summary": "getTaskVariable",
        "operationId": "getTaskVariableUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WfeVariable"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WfeExceptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "global"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Sorting": {
        "title": "Sorting",
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "order": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        }
      },
...
      "WfeExceptionResponse": {
        "title": "WfeExceptionResponse",
        "type": "object",
        "properties": {
          "date": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "AUTHENTICATION",
              "AUTHENTICATION_EXPIRED",
              "AUTHORIZATION",
              "DEFINITION_ALREADY_EXISTS",
              "DEFINITION_ARCHIVE_FORMAT",
              "DEFINITION_DOES_NOT_EXIST",
              "DEFINITION_FILE_DOES_NOT_EXIST",
              "DEFINITION_NAME_MISMATCH",
              "DEFINITION_NOT_COMPATIBLE",
              "EXECUTOR_ALREADY_EXISTS",
              "EXECUTOR_DOES_NOT_EXIST",
              "EXECUTOR_PARTICIPATES_IN_PROCESSES",
              "FILTER_FORMAT",
              "INVALID_DEFINITION",
              "OTHER",
              "PARENT_PROCESS_EXISTS",
              "PROCESS_DOES_NOT_EXIST",
              "RELATION_ALREADY_EXISTS",
              "RELATION_DOES_NOT_EXIST",
              "SUBSTITUTION_DOES_NOT_EXIST",
              "TASK_ALREADY_ACCEPTED",
              "TASK_DOES_NOT_EXIST",
              "VALIDATION",
              "WEAK_PASSWORD"
            ]
          }
        }
      },
...
      "WfePagedListFilter": {
        "title": "WfePagedListFilter",
        "type": "object",
        "properties": {
          "filters": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sortings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sorting"
            }
          },
          "variables": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "WfePagedListOfWfeTask": {
        "title": "WfePagedListOfWfeTask",
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WfeTask"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "WfeTask": {
        "title": "WfeTask",
        "type": "object",
        "properties": {
          "acquiredBySubstitution": {
            "type": "boolean"
          },
          "assignDate": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "integer",
            "format": "int64"
          },
          "deadlineDate": {
            "type": "integer",
            "format": "int64"
          },
          "deadlineWarningDate": {
            "type": "integer",
            "format": "int64"
          },
          "definitionId": {
            "type": "integer",
            "format": "int64"
          },
          "definitionName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "escalated": {
            "type": "boolean"
          },
          "firstOpen": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "multitaskIndex": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "nodeId": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/WfeExecutor"
          },
          "processHierarchyIds": {
            "type": "string"
          },
          "processId": {
            "type": "integer",
            "format": "int64"
          },
          "readOnly": {
            "type": "boolean"
          },
          "swimlaneName": {
            "type": "string"
          },
          "targetUser": {
            "$ref": "#/components/schemas/WfeUser"
          },
          "tokenId": {
            "type": "integer",
            "format": "int64"
          },
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WfeVariable"
            }
          }
        }
      },
...
      "WfeUser": {
        "title": "WfeUser",
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "department": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DELEGATION_GROUP",
              "ESCALATION_GROUP",
              "GROUP",
              "TEMPORARY_GROUP",
              "USER"
            ]
          }
        }
      },
      "WfeValidationExceptionResponse": {
        "title": "WfeValidationExceptionResponse",
        "type": "object",
        "properties": {
          "date": {
            "type": "integer",
            "format": "int64"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "globalErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "AUTHENTICATION",
              "AUTHENTICATION_EXPIRED",
              "AUTHORIZATION",
              "DEFINITION_ALREADY_EXISTS",
              "DEFINITION_ARCHIVE_FORMAT",
              "DEFINITION_DOES_NOT_EXIST",
              "DEFINITION_FILE_DOES_NOT_EXIST",
              "DEFINITION_NAME_MISMATCH",
              "DEFINITION_NOT_COMPATIBLE",
              "EXECUTOR_ALREADY_EXISTS",
              "EXECUTOR_DOES_NOT_EXIST",
              "EXECUTOR_PARTICIPATES_IN_PROCESSES",
              "FILTER_FORMAT",
              "INVALID_DEFINITION",
              "OTHER",
              "PARENT_PROCESS_EXISTS",
              "PROCESS_DOES_NOT_EXIST",
              "RELATION_ALREADY_EXISTS",
              "RELATION_DOES_NOT_EXIST",
              "SUBSTITUTION_DOES_NOT_EXIST",
              "TASK_ALREADY_ACCEPTED",
              "TASK_DOES_NOT_EXIST",
              "VALIDATION",
              "WEAK_PASSWORD"
            ]
          }
        }
      },
      "WfeVariable": {
        "title": "WfeVariable",
        "type": "object",
        "properties": {
          "format": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}